refactor(shared_kernel): promote MediaId from media module
Move MovieId/SeriesId/SeasonId/EpisodeId, the MediaId alias and
parse_media_id to the shared_kernel (ADR-018) — they are consumed
by watch_progress, collections and catalog_requests. The old path
re-exports for incremental migration; the two parse rule codes move
into a shared_kernel enum with identical string values.
