+++ src/modules/media/domain/rule_codes.py
@@ -13,8 +13,6 @@ class MediaRuleCodes(StrEnum):
     EPISODE_SERIES_MISMATCH = "EPISODE_SERIES_MISMATCH"
     EPISODE_SEASON_MISMATCH = "EPISODE_SEASON_MISMATCH"
     SEASON_SERIES_MISMATCH = "SEASON_SERIES_MISMATCH"
-    INVALID_MEDIA_ID_FORMAT = "INVALID_MEDIA_ID_FORMAT"
-    UNKNOWN_MEDIA_PREFIX = "UNKNOWN_MEDIA_PREFIX"
     INTRO_EXCEEDS_DURATION = "INTRO_EXCEEDS_DURATION"
     INTRO_DETECTION_INVALID_TRANSITION = "INTRO_DETECTION_INVALID_TRANSITION"
     MEDIA_CONFLICT_ALREADY_RESOLVED = "MEDIA_CONFLICT_ALREADY_RESOLVED"
